Move functions of _matrix_generation.py to _inputs.py
